ChatBot/ChatBot/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "openai>=1.63.0",
    "pymongo>=4.11.1",
    "python-dotenv>=1.0.1",
    "twilio>=9.4.5",
]
